Add mix.exs (#99)
This allows Elixir to compile the project without shelling
out to Rebar 3, which is useful in low memory environments.
Generally speaking, I wouldn't say this is a requirement for
projects, but given telemetry usage is widespread, it makes
sense to optimize it to different build tools.

The only downside is that the vsn can no longer be hardcoded
to "git". For this reason, we also introduce a "RELEASE.md"
file with steps to release a new version (which already
requires some extra steps due to the use of ExDoc for docs).